Adding detailed-stats option to git-linguist results
When generating Linguist statistics via the command-line currently the
only information returned is a hash of the language and it's percentage
of the repository it takes up. Sometimes it's important to also know the
file size of each language.

This adds a new option to the command=line interface called
`detailed-stats` which returns back a JSON representation of the
repository broken down by language and returning _both_ the percentage
of the repository that language comprises and the total file size in
bytes for that language. This will allow consumers an expanded ability
to extend and use the information available for their tooling.