Moving functionality to use OptionsParser
Previously, you could only choose to have a summary of the Linguist file
sizes, a breakdown including the summary and a list of files per
language, _or_ JSON output of the files per language. This has the
limitation where you cannot get all of the information in both formats.

By moving to using OptionsParser, we're able to more easily separate the
data the user is looking to get (summary or breakdown) and the format
they want the data to be returned in (CLI-friendly output or JSON).